+++ plugins/visualization-bar-chart/plugin.php
@@ -43,6 +43,11 @@ class DatawrapperPlugin_VisualizationBarChart extends DatawrapperPlugin_Visualiz
                 "absolute-scale" => array(
                     "type" => "checkbox",
                     "label" => __("Use the same scale for all columns", $id)
+                ),
+                "filter-missing-values" => array(
+                    "type" => "checkbox",
+                    "default" => true,
+                    "label" => __("Filter missing values", $id)
                 )
             ),
             "libraries" => array()